go: Change width and index from uint64 -> uint32;

+++ go/fennel/lib/counter/bucket.go
@@ -11,6 +11,8 @@ const (
 type Bucket struct {
 	Key    string
 	Window ftypes.Window
-	Width  uint64
-	Index  uint64
+	Width  uint32
+	// We set Index as `uint32` which has range [0, 4.2e^9]. With `Width` = 1, index will be in range till year ~2106
+	// assuming timestamp is in seconds since epoch, which in our case is.
+	Index  uint32
 }
