+++ go/fennel/model/counter/average.go
@@ -23,7 +23,7 @@ func NewAverage(durations []uint64) Histogram {
 	maxDuration := getMaxDuration(durations)
 	return average{
 		Durations: durations,
-		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint64{
+		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint32{
 			ftypes.Window_MINUTE: 6,
 			ftypes.Window_DAY:    1,
 		}, true},
