+++ go/fennel/model/counter/average_test.go
@@ -108,7 +108,7 @@ func TestRollingAverage_Bucketize_Valid(t *testing.T) {
 		actions.Append(d)
 		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_DAY, Index: 1, Width: 1})
 		expVals = append(expVals, value.NewList(value.Int(i), value.Int(1)))
-		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_MINUTE, Index: uint64(24*10 + i*10), Width: 6})
+		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_MINUTE, Index: uint32(24*10 + i*10), Width: 6})
 		expVals = append(expVals, value.NewList(value.Int(i), value.Int(1)))
 	}
 	buckets, values, err := Bucketize(h, actions)
