+++ go/fennel/model/counter/bucket.go
@@ -76,25 +76,26 @@ func MergeBuckets(histogram Histogram, buckets []counter.Bucket, values []value.
 // ===========================
 
 // given start, end, returns indices of [startIdx, endIdx) periods that are fully enclosed within [start, end]
-func boundary(start, end ftypes.Timestamp, period uint64) (uint64, uint64) {
-	startBoundary := (uint64(start) + period - 1) / period
-	endBoundary := uint64(end) / period
+func boundary(start, end ftypes.Timestamp, period uint32) (uint32, uint32) {
+	startBoundary := (uint32(start) + period - 1) / period
+	endBoundary := uint32(end) / period
 	return startBoundary, endBoundary
 }
 
 // trailingPartial returns the partial bucket that started within [start, end] but didn't fully finish before end
 // if there is no such bucket, ok (2nd return value) is false
-func trailingPartial(key string, start, end ftypes.Timestamp, window ftypes.Window, width uint64) (counter.Bucket, bool) {
+func trailingPartial(key string, start, end ftypes.Timestamp, window ftypes.Window, width uint32) (counter.Bucket, bool) {
 	d, err := utils.Duration(window)
 	if err != nil {
 		return counter.Bucket{}, false
 	}
+	// this will always fit in uint32
 	period := d * width
 	startBoundary, endBoundary := boundary(start, end, period)
 	if endBoundary < startBoundary {
 		return counter.Bucket{}, false
 	}
-	if endBoundary*uint64(period) == uint64(end) {
+	if endBoundary*period == uint32(end) {
 		// last bucket perfectly lines up with the end so there is no trailing partial bucket
 		return counter.Bucket{}, false
 	}
