+++ go/fennel/model/counter/max.go
@@ -31,7 +31,7 @@ func NewMax(durations []uint64) Histogram {
 	maxDuration := getMaxDuration(durations)
 	return rollingMax{
 		Durations: durations,
-		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint64{
+		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint32{
 			ftypes.Window_MINUTE: 6,
 			ftypes.Window_DAY:    1,
 		}, true},
