+++ go/fennel/model/counter/stddev_test.go
@@ -114,7 +114,7 @@ func TestStddev_Bucketize_Valid(t *testing.T) {
 			Index: 1, Width: 1})
 		expVals = append(expVals, count)
 		expected = append(expected, counter.Bucket{Key: v.String(), Window: ftypes.Window_MINUTE,
-			Index: uint64(24*10 + i*10), Width: 6})
+			Index: uint32(24*10 + i*10), Width: 6})
 		expVals = append(expVals, count)
 	}
 	buckets, vals, err := Bucketize(h, actions)
